youtube-analyser/pyproject.toml
[project]
name = "youtube-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.2.0",
    "ffmpeg-python>=0.2.0",
    "yt-dlp>=2025.8.11",
    "loguru>=0.7.0",
    "pyannote-audio>=3.1.0",
    "ollama>=0.3.0",
    "flake8>=7.3.0",
    "black>=25.1.0",
    "minio>=7.2.16",
    "python-dotenv>=1.1.1",
    "atproto>=0.0.62",
    "moviepy>=2.2.1",
    "grapheme>=0.6.0",
    "transformers>=4.55.4",
    "speechbox>=0.2.1",
    "whisperx>=3.4.2",
    "soundfile>=0.13.1",
    "qdrant-client>=1.15.1",
]
